Add split executors for panorama and world generation
Replace single worldLabsExecutor with:
- worldLabsUtils.ts: shared helpers (uploadImage, pollUntilDone, fetchWorldAssets)
- worldLabsPanoExecutor.ts: text/image/multi-image with azimuth support
- worldLabsWorldExecutor.ts: single panorama image, always "image" prompt type

Update executeNode dispatcher and workflowStore to use new executors.
